requirements.txt
# backend framework
django
djangorestframework

# database & models (Django ORM replaces SQLAlchemy)
# sqlalchemy  # removed, using Django ORM

# for interacting with an AI model via HTTP if needed
requests

# environment variable loader
python-dotenv

# required by DRF for handling file uploads
# python-multipart  # not needed with Django

# testing
pytest
